+++ tests/TestCase/Driver/MysqlTest.php
@@ -22,6 +22,7 @@
  */
 namespace MysqlBackup\Test\TestCase\Driver;
 
+use Cake\Core\Configure;
 use Cake\TestSuite\TestCase;
 use MysqlBackup\BackupTrait;
 use MysqlBackup\Driver\Mysql;
@@ -60,6 +61,12 @@ class MysqlTest extends TestCase
     {
         parent::tearDown();
 
+        //Deletes all backups
+        foreach (glob(Configure::read(MYSQL_BACKUP . '.target') . DS . '*') as $file) {
+            //@codingStandardsIgnoreLine
+            @unlink($file);
+        }
+
         unset($this->Mysql);
     }
 
@@ -93,13 +100,13 @@ class MysqlTest extends TestCase
         $method = 'getExportExecutable';
         $mysqldump = $this->getBinary('mysqldump');
 
-        $expected = $mysqldump . ' --defaults-file=%s %s | ' . $this->getBinary('bzip2') . ' > %s';
+        $expected = $mysqldump . ' --defaults-file=%s %s | ' . $this->getBinary('bzip2') . ' > %s 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Mysql, $method, ['backup.sql.bz2']));
 
-        $expected = $mysqldump . ' --defaults-file=%s %s | ' . $this->getBinary('gzip') . ' > %s';
+        $expected = $mysqldump . ' --defaults-file=%s %s | ' . $this->getBinary('gzip') . ' > %s 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Mysql, $method, ['backup.sql.gz']));
 
-        $expected = $mysqldump . ' --defaults-file=%s %s > %s';
+        $expected = $mysqldump . ' --defaults-file=%s %s > %s 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Mysql, $method, ['backup.sql']));
     }
 
@@ -141,13 +148,13 @@ class MysqlTest extends TestCase
         $method = 'getImportExecutable';
         $mysql = $this->getBinary('mysql');
 
-        $expected = $this->getBinary('bzip2') . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s';
+        $expected = $this->getBinary('bzip2') . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Mysql, $method, ['backup.sql.bz2']));
 
-        $expected = $this->getBinary('gzip') . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s';
+        $expected = $this->getBinary('gzip') . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Mysql, $method, ['backup.sql.gz']));
 
-        $expected = 'cat %s | ' . $mysql . ' --defaults-extra-file=%s %s';
+        $expected = 'cat %s | ' . $mysql . ' --defaults-extra-file=%s %s 2>/dev/null';
         $this->assertEquals($expected, $this->invokeMethod($this->Mysql, $method, ['backup.sql']));
     }
 
@@ -184,5 +191,66 @@ class MysqlTest extends TestCase
     public function testGetValidCompressions()
     {
         $this->assertEquals(['sql.bz2' => 'bzip2', 'sql.gz' => 'gzip', 'sql' => false], $this->Mysql->getValidCompressions());
+    }
+
+    /**
+     * Test for `export()` method
+     * @test
+     */
+    public function testExport()
+    {
+        $filename = Configure::read(MYSQL_BACKUP . '.target') . DS . 'test.sql';
+        $export = $this->Mysql->export($filename);
+
+        $this->assertTrue($export);
+        $this->assertFileExists(Configure::read(MYSQL_BACKUP . '.target') . DS . 'test.sql');
+    }
+
+    /**
+     * Test for `export()` method, with a failure by mysqldump
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage mysqldump failed with exit code `2`
+     * @test
+     */
+    public function testExportMysqldumpFailure()
+    {
+        //Sets a no existing database
+        $connection = $this->getProperty($this->Mysql, 'connection');
+        $connection['database'] = 'noExisting';
+        $this->setProperty($this->Mysql, 'connection', $connection);
+
+        $this->Mysql->export(Configure::read(MYSQL_BACKUP . '.target') . DS . 'test.sql');
+    }
+
+    /**
+     * Test for `import()` method
+     * @test
+     */
+    public function testImport()
+    {
+        $filename = Configure::read(MYSQL_BACKUP . '.target') . DS . 'test.sql';
+        $this->Mysql->export($filename);
+
+        $import = $this->Mysql->import($filename);
+        $this->assertTrue($import);
+    }
+
+    /**
+     * Test for `import()` method, with a failure by mysql
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage mysql failed with exit code `1`
+     * @test
+     */
+    public function testImportMysqlFailure()
+    {
+        $filename = Configure::read(MYSQL_BACKUP . '.target') . DS . 'test.sql';
+        $this->Mysql->export($filename);
+
+        //Sets a no existing database
+        $connection = $this->getProperty($this->Mysql, 'connection');
+        $connection['database'] = 'noExisting';
+        $this->setProperty($this->Mysql, 'connection', $connection);
+
+        $this->Mysql->import($filename);
     }
 }