+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## 2.0 branch
 ### 2.0.0
 * the code has been completely rewritten to work with drivers, so it can also
-    work with other database engines.
+    work with other database engines;
+* checks the return status code when it runs `mysql` and `mysqldump` from the
+    command line.
 
 # 1.x branch
 ## 1.1 branch